+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/jvm/JavaDevelopmentKit.kt
@@ -6,4 +6,5 @@ interface JavaDevelopmentKit {
     fun install(connection: SshConnection)
     fun use(): String
     fun command(options: String): String
+    val jstatMonitoring: Jstat
 }