+++ src/test/kotlin/com/atlassian/performance/tools/infrastructure/api/jvm/OracleJdkIT.kt
@@ -1,0 +1,11 @@
+package com.atlassian.performance.tools.infrastructure.api.jvm
+
+import org.junit.Test
+
+class OracleJdkIT {
+
+    @Test
+    fun shouldSupportJstat() {
+        JstatSupport(OracleJDK()).shouldSupportJstat()
+    }
+}